Refactor: 확장한 Date 클래스를 Data.swift 파일에 옮김
	이유: 이후에 우유 유통기한에도 확장한 Date를 쓸거라서
	  잘 보이기 위해 새로 파일을 만들었다.
