+++ VendingMachineApp/VendingMachineApp/Models/Date.swift
@@ -1,0 +1,23 @@
+//
+//  Date.swift
+//  VendingMachineApp
+//
+//  Created by kimdo2297 on 2020/03/09.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import Foundation
+
+extension Date {
+    
+    static let dateFormatter : DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyyMMdd"
+        return formatter
+    }()
+
+    var description: String {
+        return Date.dateFormatter.string(from: self)
+    }
+    
+}
